+++ atcodertools/codegen/codegen_modules/cpp.py
@@ -161,6 +161,9 @@ class NoPredictionResultGiven(Exception):
 
 
 def main(args: CodeGenArgs) -> Dict[str, Any]:
+    """
+    :return: A dictionary of the parameters passed to template.
+    """
     code_parameters = CppCodeGenerator(args.format, args.config).generate_parameters()
     return {
         **code_parameters,
