+++ atcodertools/codegen/codegen_modules/java.py
@@ -1,7 +1,7 @@
 from typing import Dict, Any
 
 from atcodertools.codegen.models.codegen_args import CodeGenArgs
-from atcodertools.codegen.langs.cpp import CppCodeGenerator
+from atcodertools.codegen.codegen_modules.cpp import CppCodeGenerator
 from atcodertools.fmtprediction.models.type import Type
 from atcodertools.fmtprediction.models.variable import Variable
 
@@ -68,6 +68,9 @@ class JavaCodeGenerator(CppCodeGenerator):
 
 
 def main(args: CodeGenArgs) -> Dict[str, Any]:
+    """
+    :return: A dictionary of the parameters passed to template.
+    """
     code_parameters = JavaCodeGenerator(args.format, args.config).generate_parameters()
     return {
         **code_parameters,
